% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kuromoji.R
\name{kuromoji}
\alias{kuromoji}
\title{Call kuromoji tokenizer}
\usage{
kuromoji(chr, split = TRUE)
}
\arguments{
\item{chr}{Character vector to be tokenized.}

\item{split}{Logical.If true (by default), the function splits character vector
into sentences using \code{stringi::stri_split_boudaries(type = "sentence")}
before analyzing them.}
}
\value{
List.
}
\description{
Call kuromoji tokenizer
}
