% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pack.R
\name{pack}
\alias{pack}
\title{Pack prettified output}
\usage{
pack(df, pull = "token", .collapse = " ")
}
\arguments{
\item{df}{Output of \code{tangela::pack}.}

\item{pull}{Column name to be packed into data.frame. Default value is \code{token}.}

\item{.collapse}{This argument is passed to \code{stringi::stri_c()}.}
}
\value{
data.frame
}
\description{
Pack prettified output
}
