% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{rebuild_tokenizer}
\alias{rebuild_tokenizer}
\title{Initialize kuromoji tokenizer}
\usage{
rebuild_tokenizer(user_dic = "")
}
\arguments{
\item{user_dic}{file path to a user dictionary if any.}
}
\value{
The stored kuromoji tokenizer instance is returned invisibly.
}
\description{
Initialize kuromoji tokenizer
}
