% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tangela-package.R
\docType{package}
\name{tangela-package}
\alias{tangela}
\alias{tangela-package}
\title{tangela: rJava Interface to Kuromoji}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

An rJava wrapper for atilika/kuromoji (v0.7.7). This package will work fine, but it is too slow to be used in production.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/paithiov909/tangela}
  \item Report bugs at \url{https://github.com/paithiov909/tangela/issues}
}

}
\author{
\strong{Maintainer}: Akiru Kato \email{paithiov909@gmail.com}

}
\keyword{internal}
