% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tangela.R
\docType{package}
\name{tangela}
\alias{tangela}
\alias{tangela-package}
\title{tangela: rJava Interface to Kuromoji}
\description{
An rJava interface to atilika/kuromoji (v0.7.7).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/paithiov909/tangela}
  \item Report bugs at \url{https://github.com/paithiov909/tangela/issues}
}

}
\author{
\strong{Maintainer}: Kato Akiru \email{paithiov909@gmail.com}

}
