% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kuromoji.R
\name{kuromoji}
\alias{kuromoji}
\title{Call Kuromoji Tokenizer}
\usage{
kuromoji(str, userDic = "")
}
\arguments{
\item{str}{String to be tokenized.}

\item{userDic}{String}
}
\value{
a list
}
\description{
Call Kuromoji Tokenizer
}
