% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kuromoji.R
\name{kuromoji}
\alias{kuromoji}
\title{Call kuromoji tokenizer}
\usage{
kuromoji(chr)
}
\arguments{
\item{chr}{Character vector to be tokenized.}
}
\value{
A data.frame
}
\description{
Call kuromoji tokenizer
}
