% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{rebuild_tokenizer}
\alias{rebuild_tokenizer}
\title{Initialize Kuromoji Tokenizer}
\usage{
rebuild_tokenizer(user_dic = "")
}
\arguments{
\item{user_dic}{character scalar. file path to user dictionary if any.}
}
\value{
returns kuromoji tokenizer instance invisibly.
}
\description{
Initialize Kuromoji Tokenizer
}
